Fix antags
