+++ Content.Server/Administration/Systems/AdminVerbSystem.Antags.cs
@@ -1,6 +1,7 @@
 using Content.Server.Administration.Commands;
 using Content.Server.Antag;
 using Content.Server.Backmen.GameTicking.Rules.Components;
+using Content.Server.Backmen.Vampiric;
 using Content.Server.GameTicking.Rules.Components;
 using Content.Server.Zombies;
 using Content.Shared.Administration;
@@ -81,6 +82,23 @@ public sealed partial class AdminVerbSystem
             Message = Loc.GetString("admin-verb-text-make-blob"),
         };
         args.Verbs.Add(blobAntag);
+
+        Verb vampireAntag = new()
+        {
+            Text = Loc.GetString("admin-verb-text-make-vampire"),
+            Category = VerbCategory.Antag,
+            Icon = new SpriteSpecifier.Texture(new ("/Textures/Backmen/Icons/verbiconfangs.png")),
+            Act = () =>
+            {
+                if (!HasComp<ActorComponent>(args.Target))
+                    return;
+
+                EntityManager.System<BloodSuckerSystem>().ForceMakeVampire(args.Target);
+            },
+            Impact = LogImpact.High,
+            Message = Loc.GetString("admin-verb-text-make-vampire"),
+        };
+        args.Verbs.Add(vampireAntag);
 
         Verb fleshLeaderCultist = new()
         {
