+++ Content.Server/Backmen/SpecForces/SpecForcesSystem.cs
@@ -16,7 +16,6 @@ using System.Threading;
 using Content.Server.Actions;
 using Content.Server.Backmen.Blob;
 using Content.Server.Backmen.Blob.Components;
-using Content.Server.Backmen.Blob.Rule;
 using Content.Server.Backmen.GameTicking.Rules.Components;
 using Content.Server.Ghost.Roles.Components;
 using Content.Server.RandomMetadata;
