+++ Content.Server/Backmen/StationEvents/Events/BlobSpawn.cs
@@ -68,7 +68,7 @@ public sealed class BlobSpawnRule : StationEventSystem<BlobSpawnRuleComponent>
         }
 
         // start blob rule incase it isn't, for the sweet greentext
-        GameTicker.StartGameRule("Blob");
+        GameTicker.StartGameRule("BlobRule");
     }
 
     // Because GameRule spawns just a GhostRoleSpawner, we can't just remove components
