--- Content.Server/Backmen/Vampiric/Goal/BloodsuckerObjectiveComponent.cs
@@ -1,7 +1,0 @@
-﻿namespace Content.Server.Backmen.Vampiric.Goal;
-
-[RegisterComponent]
-public sealed partial class BloodsuckerObjectiveComponent : Component
-{
-
-}
