+++ Content.Server/Backmen/Vampiric/Objective/BloodsuckerConvertConditionComponent.cs
@@ -1,4 +1,4 @@
-﻿namespace Content.Server.Backmen.Vampiric.Role;
+﻿namespace Content.Server.Backmen.Vampiric.Objective;
 
 [RegisterComponent]
 public sealed partial class BloodsuckerConvertConditionComponent : Component
