+++ Content.Server/Backmen/Vampiric/Rule/BloodsuckerRuleComponent.cs
@@ -1,6 +1,4 @@
-﻿using Robust.Shared.Audio;
-
-namespace Content.Server.Backmen.Vampiric;
+﻿namespace Content.Server.Backmen.Vampiric.Rule;
 
 [RegisterComponent]
 public sealed partial class BloodsuckerRuleComponent : Component
@@ -8,18 +6,4 @@ public sealed partial class BloodsuckerRuleComponent : Component
     public readonly Dictionary<string, EntityUid> Elders = new();
 
     public int TotalBloodsuckers = 0;
-
-    public List<string> SpeciesWhitelist = new()
-    {
-        "Human",
-        "Reptilian",
-        "Dwarf",
-        "Oni",
-        "Vox",
-        "HumanoidFoxes",
-        "Harpy"
-    };
-
-    [DataField]
-    public SoundSpecifier InitialInfectedSound = new SoundPathSpecifier("/Audio/Backmen/Ambience/Antag/vampier_start.ogg");
 }
