+++ Content.Server/Backmen/Vampiric/Rule/BloodsuckerRuleSystem.cs
@@ -1,192 +1,129 @@
-﻿using System.Linq;
-using Content.Server.Antag;
+﻿using Content.Server.Antag;
+using Content.Server.Backmen.Vampiric.Objective;
 using Content.Server.Backmen.Vampiric.Role;
-using Content.Server.Bible.Components;
-using Content.Server.Chat.Managers;
 using Content.Server.GameTicking;
 using Content.Server.GameTicking.Rules;
-using Content.Server.GameTicking.Rules.Components;
 using Content.Server.Mind;
-using Content.Server.Objectives;
-using Content.Shared.Antag;
-using Content.Shared.Backmen.CCVar;
-using Content.Shared.Backmen.Vampiric;
 using Content.Shared.GameTicking.Components;
-using Content.Shared.Humanoid;
 using Content.Shared.Mind;
-using Content.Shared.Mobs.Components;
-using Content.Shared.Preferences;
-using Content.Shared.Roles;
-using Robust.Server.Placement;
-using Robust.Server.Player;
-using Robust.Shared.Configuration;
-using Robust.Shared.Player;
-using Robust.Shared.Prototypes;
-using Robust.Shared.Random;
-using Robust.Shared.Timing;
 
-namespace Content.Server.Backmen.Vampiric;
+namespace Content.Server.Backmen.Vampiric.Rule;
 
 public sealed class BloodsuckerRuleSystem : GameRuleSystem<BloodsuckerRuleComponent>
 {
-
-    private ISawmill _sawmill = default!;
-
-    private int PlayersPerBloodsucker => _cfg.GetCVar(CCVars.BloodsuckerPlayersPerBloodsucker);
-    private int MaxBloodsuckers => _cfg.GetCVar(CCVars.BloodsuckerMaxPerBloodsucker);
-
-    [Dependency] private readonly IConfigurationManager _cfg = default!;
-    //[Dependency] private readonly IPrototypeManager _prototypeManager = default!;
-    //[Dependency] private readonly IChatManager _chatManager = default!;
-    [Dependency] private readonly OldAntagSelectionSystem _antagSelection = default!;
     [Dependency] private readonly BloodSuckerSystem _bloodSuckerSystem = default!;
-    [Dependency] private readonly IRobustRandom _random = default!;
     [Dependency] private readonly MindSystem _mindSystem = default!;
-    [Dependency] private readonly GameTicker _gameTicker = default!;
-    [Dependency] private readonly IPlayerManager _playerManager = default!;
 
     public override void Initialize()
     {
         base.Initialize();
 
-        _sawmill = Logger.GetSawmill("preset");
-
-        SubscribeLocalEvent<RulePlayerJobsAssignedEvent>(OnPlayersSpawned);
-        SubscribeLocalEvent<PlayerSpawnCompleteEvent>(HandleLatejoin);
+        SubscribeLocalEvent<BloodsuckerRuleComponent, AfterAntagEntitySelectedEvent>(AfterAntagSelected);
     }
 
-    [ValidatePrototypeId<EntityPrototype>]
-    private const string VampireObjective = "VampireObjective";
-
-    protected override void Added(EntityUid uid, BloodsuckerRuleComponent component, GameRuleComponent gameRule, GameRuleAddedEvent args)
+    private void AfterAntagSelected(EntityUid uid, BloodsuckerRuleComponent component, AfterAntagEntitySelectedEvent args)
     {
-        base.Added(uid, component, gameRule, args);
-
-        if (_gameTicker.RunLevel == GameRunLevel.InRound)
-        {
-            _gameTicker.StartGameRule(VampireObjective);
-        }
-        else
-        {
-            _gameTicker.AddGameRule(VampireObjective);
-        }
+        _bloodSuckerSystem.MakeVampire(args.EntityUid, true);
     }
 
-    private void HandleLatejoin(PlayerSpawnCompleteEvent ev)
+    protected override void AppendRoundEndText(EntityUid uid,
+        BloodsuckerRuleComponent component,
+        GameRuleComponent gameRule,
+        ref RoundEndTextAppendEvent ev)
     {
-        var query = QueryActiveRules();
-        while (query.MoveNext(out _, out var vpmRule, out _))
+        var skip = new List<EntityUid>();
+        var query = AllEntityQuery<BloodsuckerRuleComponent>();
+        while (query.MoveNext(out var vampRule))
         {
-            if (vpmRule.TotalBloodsuckers >= MaxBloodsuckers)
-                continue;
+            if(vampRule.Elders.Count == 0)
+                break;
+            ev.AddLine(Loc.GetString("vampire-elder"));
 
-            if (!ev.LateJoin)
-                continue;
-
-            var whitelistSpecies = vpmRule.SpeciesWhitelist;
-
-            if (!_antagSelection.IsPlayerEligible(ev.Player, Bloodsucker, acceptableAntags: AntagAcceptability.NotExclusive,
-                    allowNonHumanoids: false,
-                    customExcludeCondition: ent =>
-                    {
-                        if (HasComp<BibleUserComponent>(ent))
-                        {
-                            return true;
-                        }
-
-                        return TryComp<HumanoidAppearanceComponent>(ent, out var humanoidAppearanceComponent) &&
-                               !whitelistSpecies.Contains(humanoidAppearanceComponent.Species.Id);
-                    }))
-                continue;
-
-            // the nth player we adjust our probabilities around
-            var target = PlayersPerBloodsucker * vpmRule.TotalBloodsuckers + 1;
-
-            var chance = 1f / PlayersPerBloodsucker;
-
-            // If we have too many traitors, divide by how many players below target for next traitor we are.
-            if (ev.JoinOrder < target)
+            foreach (var player in vampRule.Elders)
             {
-                chance /= (target - ev.JoinOrder);
-            }
-            else // Tick up towards 100% chance.
-            {
-                chance *= ((ev.JoinOrder + 1) - target);
-            }
+                skip.Add(player.Value);
+                var role = CompOrNull<VampireRoleComponent>(player.Value);
+                var count = role?.Converted ?? 0;
+                var blood = role?.Drink ?? 0;
+                var countGoal = 0;
+                var bloodGoal = 0f;
 
-            if (chance > 1)
-                chance = 1;
-
-            // Now that we've calculated our chance, roll and make them a traitor if we roll under.
-            // You get one shot.
-            if (_random.Prob(chance) && ev.Player.AttachedEntity.HasValue)
-            {
-                _bloodSuckerSystem.ConvertToVampire(ev.Player.AttachedEntity.Value);
-                vpmRule.TotalBloodsuckers++;
-                if (_mindSystem.TryGetMind(ev.Player, out var mindId, out _))
+                var mind = CompOrNull<MindComponent>(player.Value);
+                if (_mindSystem.TryGetObjectiveComp<BloodsuckerDrinkConditionComponent>(player.Value, out var obj1, mind))
                 {
-                    vpmRule.Elders.Add(MetaData(ev.Player.AttachedEntity.Value).EntityName,mindId);
+                    bloodGoal = obj1.Goal;
                 }
-                _antagSelection.SendBriefing(ev.Player, Loc.GetString("vampire-role-greeting"), Color.Plum, vpmRule.InitialInfectedSound);
-            }
-        }
-    }
-
-    private void OnPlayersSpawned(RulePlayerJobsAssignedEvent ev)
-    {
-        var query = QueryActiveRules();
-        while (query.MoveNext(out var uid, out _, out var comp, out var gameRule))
-        {
-            //Get all players eligible for this role, allow selecting existing antags
-            //TO DO: When voxes specifies are added, increase their chance of becoming a thief by 4 times >:)
-            var whitelistSpecies = comp.SpeciesWhitelist;
-            var eligiblePlayers = _antagSelection.GetEligiblePlayers(
-                ev.Players, Bloodsucker,
-                acceptableAntags: AntagAcceptability.NotExclusive,
-                allowNonHumanoids: false,
-                customExcludeCondition: ent =>
+                if (_mindSystem.TryGetObjectiveComp<BloodsuckerConvertConditionComponent>(player.Value, out var obj2, mind))
                 {
-                    if (HasComp<BibleUserComponent>(ent))
-                    {
-                        return true;
-                    }
+                    countGoal = obj2.Goal;
+                }
 
-                    return TryComp<HumanoidAppearanceComponent>(ent, out var humanoidAppearanceComponent) &&
-                           !whitelistSpecies.Contains(humanoidAppearanceComponent.Species.Id);
-                });
+                _mindSystem.TryGetSession(player.Value, out var session);
+                var username = session?.Name;
+                if (username != null)
+                {
+                    ev.AddLine(Loc.GetString("endgame-vamp-name-user", ("name", player.Key), ("username", username)));
+                }
+                else
+                {
+                    ev.AddLine(Loc.GetString("endgame-vamp-name", ("name", player.Key)));
+                }
+                ev.AddLine(Loc.GetString("endgame-vamp-conv",
+                    ("count", count),
+                    ("goal", countGoal)));
+                ev.AddLine(Loc.GetString("endgame-vamp-drink",
+                    ("count", blood),
+                    ("goal", bloodGoal)));
+            }
 
-            //Abort if there are none
-            if (eligiblePlayers.Count == 0)
+            ev.AddLine("");
+        }
+
+        var isAddLine = true;
+
+        var q = EntityQueryEnumerator<MindComponent,VampireRoleComponent>();
+        while (q.MoveNext(out var mindId,out var mind, out var role))
+        {
+            if (skip.Contains(mindId))
             {
-                //Log.Warning($"No eligible thieves found, ending game rule {ToPrettyString(uid):rule}");
-                //GameTicker.EndGameRule(uid, gameRule);
                 continue;
             }
 
-            var initialInfectedCount = _antagSelection.CalculateAntagCount(_playerManager.PlayerCount, PlayersPerBloodsucker, MaxBloodsuckers);
-
-            //Select our theives
-            var thieves = _antagSelection.ChooseAntags(initialInfectedCount, eligiblePlayers);
-
-            DoVampirStart(thieves, comp);
-
-            _antagSelection.SendBriefing(thieves, Loc.GetString("vampire-role-greeting"), Color.Plum, comp.InitialInfectedSound);
-        }
-    }
-
-    [ValidatePrototypeId<AntagPrototype>]
-    private const string Bloodsucker = "Bloodsucker";
-
-    private void DoVampirStart(List<EntityUid> startCandidates, BloodsuckerRuleComponent vpmRule)
-    {
-        foreach (var traitor in startCandidates)
-        {
-            _bloodSuckerSystem.ConvertToVampire(traitor);
-            vpmRule.TotalBloodsuckers++;
-            if (_mindSystem.TryGetMind(traitor, out var mindId, out _))
+            if (isAddLine)
             {
-                vpmRule.Elders.Add(MetaData(traitor).EntityName,mindId);
+                ev.AddLine(Loc.GetString("vampire-bitten"));
+                isAddLine = false;
             }
+            var count = role?.Converted ?? 0;
+            var blood = role?.Drink ?? 0;
+            var countGoal = 0;
+            var bloodGoal = 0f;
+
+            if (_mindSystem.TryGetObjectiveComp<BloodsuckerDrinkConditionComponent>(mindId, out var obj1,mind))
+            {
+                bloodGoal = obj1.Goal;
+            }
+            if (_mindSystem.TryGetObjectiveComp<BloodsuckerConvertConditionComponent>(mindId, out var obj2,mind))
+            {
+                countGoal = obj2.Goal;
+            }
+
+            _mindSystem.TryGetSession(mindId, out var session);
+            var username = session?.Name;
+            if (username != null)
+            {
+                ev.AddLine(Loc.GetString("endgame-vamp-name-user", ("name", mind.CharacterName ?? "-"), ("username", username)));
+            }
+            else
+            {
+                ev.AddLine(Loc.GetString("endgame-vamp-name", ("name", mind.CharacterName ?? "-")));
+            }
+            ev.AddLine(Loc.GetString("endgame-vamp-conv",
+                ("count", count),
+                ("goal", countGoal)));
+            ev.AddLine(Loc.GetString("endgame-vamp-drink",
+                ("count", blood),
+                ("goal", bloodGoal)));
         }
     }
 }
