+++ Resources/Locale/ru-RU/administration/antag.ftl
@@ -5,7 +5,8 @@ admin-verb-make-zombie = Сделать цель зомби.
 admin-verb-make-nuclear-operative = Сделать цель одиноким Ядерным оперативником.
 admin-verb-make-pirate = Сделать цель пиратом\капером. Учтите, что это не меняет игровой режим.
 admin-verb-make-eviltwin = Сделать цели злого близнеца.
-admin-verb-text-make-blob = Сделать блобом
+admin-verb-text-make-blob = Сделать цель блобом.
+admin-verb-text-make-vampire = Сделать цель вампиром.
 admin-verb-text-make-flesh-leader-cultist = Сделать лидером культа плоти
 admin-verb-text-make-flesh-cultist = Сделать культистом плоти
 admin-verb-make-space-ninja = Сделать цель космическим ниндзя.
