+++ Resources/Prototypes/_Backmen/Objectives/blobObjectives.yml
@@ -12,7 +12,7 @@
       state: blob_nuke_overlay
   - type: RoleRequirement
     roles:
-      components:
+      mindRoles:
       - BlobRole
 
 - type: entity
