+++ Resources/Prototypes/_Backmen/Objectives/changeling.yml
@@ -8,7 +8,7 @@
     issuer: objective-issuer-hivemind
   - type: RoleRequirement
     roles:
-      components:
+      mindRoles:
       - ChangelingRole
 
 - type: entity
