+++ Resources/Prototypes/_Backmen/Objectives/superPsi.yml
@@ -7,7 +7,7 @@
     issuer: objective-issuer-superpsi
   - type: RoleRequirement
     roles:
-      components:
+      mindRoles:
       - SuperPsiRole
 
 - type: entity
