+++ Resources/Prototypes/_Backmen/Objectives/traitorGorlex.yml
@@ -5,7 +5,7 @@
   components:
   - type: RoleRequirement
     roles:
-      components:
+      mindRoles:
       - TraitorRole
   - type: Objective
     issuer: issuer-Gorlex-Marauders
