+++ Resources/Prototypes/_Backmen/Objectives/traitorMI13.yml
@@ -5,7 +5,7 @@
   components:
   - type: RoleRequirement
     roles:
-      components:
+      mindRoles:
       - TraitorRole
   - type: Objective
     issuer: issuer-MI13
